requirements.txt
asgiref==3.8.1
Django==4.2.21
django-role-permissions==3.2.0
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
